src/requirements.txt
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
flask-mongoengine==1.0.0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.23.1
Pillow==8.2.0
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.2
pymongo==3.11.4
PyYAML==5.3.1
Werkzeug==1.0.1
WTForms==2.3.3
flask-prometheus-metrics==1.0.0
